hack/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
greenlet==3.0.3
    # via playwright
playwright==1.44.0
    # via -r requirements.in
pyee==11.1.0
    # via playwright
pyyaml==6.0.1
    # via -r requirements.in
typing-extensions==4.12.0
    # via pyee
